requirements.txt
Flask==3.0.2
gunicorn==22.0.0
openpyxl==3.1.5
requests==2.31.0
python-dotenv==1.0.1
python-docx==1.1.2
pypdf==5.1.0
Pillow==11.0.0
reportlab==4.2.5
pytesseract==0.3.13
urllib3==1.26.18
charset-normalizer==3.3.2